+++ config/dev.exs
@@ -1,7 +1,7 @@
 use Mix.Config
 
 config :ex_money,
-  exchange_rate_service: true,
+  exchange_rate_service: false,
   open_exchange_rates_app_id: {:system, "OPEN_EXCHANGE_RATES_APP_ID"},
   open_exchange_rates_retrieve_every: 300_000,
   callback_module: Money.ExchangeRates.Callback,